// src/convex/_generated/api.ts
// This file will be auto-generated by Convex when the server is running

export const api = {
  projects: {
    getAll: "projects:getAll",
    getById: "projects:getById",
    create: "projects:create",
    update: "projects:update",
    remove: "projects:remove",
  },
  messages: {
    getAll: "messages:getAll",
    getById: "messages:getById",
    create: "messages:create",
    markAsRead: "messages:markAsRead",
    remove: "messages:remove",
  },
  contacts: {
    getAll: "contacts:getAll",
    getById: "contacts:getById",
    create: "contacts:create",
    update: "contacts:update",
    remove: "contacts:remove",
  },
  files: {
    generateUploadUrl: "files:generateUploadUrl",
    createFile: "files:createFile",
    getAll: "files:getAll",
    getByProject: "files:getByProject",
    getUrl: "files:getUrl",
    getUrlMutation: "files:getUrlMutation",
    remove: "files:remove",
  }
};